refactor(codegen): extract AionType→LLVM type lowering into types.rs (#113 phase 2)
src/codegen/compiler.rs was 3 699 lines after phase 1 (#119). Per the
phase tracker in #113, extract the type-emission functions into the new
sibling submodule src/codegen/types.rs (163 lines):

- aion_type_to_llvm: Aion type-name string -> BasicTypeEnum
- parse_array_type_name: [T; N] -> (elem_llvm, count). #54
- ensure_tuple_type: register anonymous tuple struct, forward-ref safe. #53
- coerce_int_width: i*/u* sign/zero-extend + trunc for let binding. #52
- type_to_llvm: analysis Type -> BasicTypeEnum; owns the box-representation
  invariant for structs/enums/tuples/arrays (docs/architecture.md).

Visibility follows the phase-1 convention: moved methods become
pub(in crate::codegen) so compiler.rs calls self.aion_type_to_llvm(...),
self.parse_array_type_name(...), self.ensure_tuple_type(...),
self.coerce_int_width(...), self.type_to_llvm(...) unchanged. The
Compiler public API is unchanged — crate-internal visibility only.

Behaviour-preserving code motion: 101 integration tests green, snapshots
byte-identical before/after, cargo clippy --all-targets -- -D warnings
clean (fixed an unused BasicType import surfaced during the move),
cargo fmt --check clean, scripts/docs-check.sh clean.

compiler.rs: 3 699 -> 3 561 lines (-138). #113 stays open for phases 3+
(generics.rs / control_flow.rs / lvalues.rs / expressions.rs /
type_helpers.rs).
