+++ src/codegen/mod.rs
@@ -1,5 +1,6 @@
 pub mod compiler;
 mod intrinsics;
 pub mod transpiler;
+mod types;
 
 pub use compiler::Compiler;
